requirements.txt
absl-py==2.1.0
accelerate==0.30.0
adagio==0.2.4
aiohttp==3.9.1
aiosignal==1.3.1
antlr4-python3-runtime==4.11.1
anyio==4.2.0
appdirs==1.4.4
appnope==0.1.3
arch==6.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.14.0
beautifulsoup4==4.12.2
bitarray==2.9.0
bleach==6.1.0
catboost==1.2.2
category-encoders==2.6.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmake==3.29.5.1
cmdstanpy==1.2.2
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cramjam==2.7.0
cycler==0.12.1
Cython==3.0.7
cytoolz==0.12.2
datasets==2.19.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
docker-pycreds==0.4.0
e2cnn==0.2.3
et-xmlfile==1.1.0
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.7.0
eth-keys==0.4.0
eth-rlp==1.0.0
eth-typing==3.5.2
eth-utils==2.3.1
evaluate==0.4.2
evmdasm==0.1.10
executing==2.0.1
fastjsonschema==2.19.0
fastparquet==2023.10.1
fft-conv-pytorch==1.2.0
filelock==3.13.1
fonttools==4.46.0
fqdn==1.5.1
frozendict==2.4.4
frozenlist==1.4.1
fs==2.4.16
fsspec==2023.12.2
fugue==0.8.7
fugue-sql-antlr==0.2.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.42
graphviz==0.20.1
grpcio==1.62.0
h11==0.14.0
hexbytes==0.3.1
holidays==0.45
html5lib==1.1
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.23.0
hyperopt==0.2.7
idna==3.6
imbalanced-learn==0.11.0
imblearn==0.0
importlib_resources==6.4.0
iniconfig==2.0.0
investpy==1.0.8
ipykernel==6.27.1
ipython==8.18.1
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.1
jupyter_server==2.12.1
jupyter_server_terminals==0.5.0
jupyterlab==4.1.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lightgbm==4.3.0
lightning-utilities==0.11.2
llvmlite==0.41.1
lru-dict==1.2.0
lxml==5.2.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
multitasking==0.0.11
nbclient==0.9.0
nbconvert==7.13.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
notebook==7.1.1
notebook_shim==0.2.3
numba==0.58.1
numpy==1.26.2
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
overrides==7.4.0
packaging==23.2
pandas==2.1.4
pandas-datareader==0.10.0
pandocfilters==1.5.0
parsimonious==0.9.0
parso==0.8.3
patsy==0.5.6
peewee==3.17.5
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
plotly==5.18.0
pluggy==1.4.0
pmdarima==2.0.4
portalocker==2.8.2
progressbar2==4.3.2
prometheus-client==0.19.0
prompt-toolkit==3.0.43
prophet==1.1.5
protobuf==4.25.1
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==15.0.2
pyarrow-hotfix==0.6
pycparser==2.21
pycryptodome==3.19.0
pygam==0.9.0
pygame==2.5.2
Pygments==2.17.2
pyparsing==3.1.1
pyspark==3.5.1
pytest==8.0.0
python-dateutil==2.8.2
python-json-logger==2.0.7
python-utils==3.8.2
pytorch-lightning==2.2.4
pytz==2023.3.post1
pyunormalize==15.1.0
PyYAML==6.0.1
pyzmq==25.1.2
qpd==0.4.4
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rlp==4.0.0
rpds-py==0.15.2
sacrebleu==2.4.2
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.12.0
seaborn==0.13.0
Send2Trash==1.8.2
sentry-sdk==1.41.0
setproctitle==1.3.3
shap==0.44.0
six==1.16.0
slicer==0.0.7
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
sqlglot==23.3.0
stack-data==0.6.3
stanio==0.5.0
statsforecast==1.7.3
statsmodels==0.14.1
stumpy==1.12.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
terminado==0.18.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.19.1
toolz==0.12.0
torch==2.2.2
torchaudio==2.2.2
torchmetrics==1.3.2
torchvision==0.17.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
transformers==4.40.1
triad==0.9.6
tsfresh==0.20.2
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.3
Unidecode==1.3.8
uri-template==1.3.0
urllib3==2.1.0
utilsforecast==0.1.2
validators==0.28.3
wandb==0.16.4
wcwidth==0.2.12
web3==6.12.0
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
widgetsnbextension==4.0.10
xgboost==2.0.3
xxhash==3.4.1
yarl==1.9.4
yfinance==0.2.40
